+++ src/index.ts
@@ -33,6 +33,7 @@ export * from "./stock";
 export * from "./store";
 export * from "./storefrontEvent";
 export * from "./subscription";
+export * from "./support";
 export * from "./supplier";
 export * from "./user";
 export * from "./userActivity";